requirements.txt
flask
gunicorn
Pillow
requests
googletrans==4.0.0-rc1
flask-cors